pyproject.toml
[tool.poetry]
name = "transfer-messages"
version = "0.1.0"
description = ""
authors = ["DimaSimakov <dimarodeo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
pyrogram = "^2.0.106"
tgcrypto = "^1.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
